// src/common-pages/proxy/proxy.type.ts
export interface AgentInfo {
  // 代理申请状态：0未申请，1已申请审核中、2申请通过、3申请拒绝
  agentApplyStatus: number;
  // 代理等级
  agentLevel: number;
  // 邀請码
  inviteCode: string;
  registerDate: string;
  todayActiveUserCount: number;
  todayBetAmount: string;
  todayRechargeUserCount: number;
  // 今日佣金金额
  todayCommissionAmount: string;
  todayNewUserCount: number;
  todayRechargeAmount: string;

  totalBetAmount: string;
  totalCommissionAmount: string;
  totalRechargeAmount: string;
  totalUserCount: number;

  // 用户id（原样展示）
  userId: number;
  // 用户手机号
  userPhone: string;
  // 用户头像
  userAvatar?: string;
}